Reorganize tests: separate strategy tests from group_event_loop
- Create test_strategy.cpp for runtime strategy tests (Hybrid wait path)
- Rename test_group_event_loop_constexpr.cpp to test_strategy_constexpr.cpp
- Remove Hybrid tests from test_group_event_loop.cpp
- Update CMakeLists.txt to include new test files

This separates concerns: strategy behavior tests are now in their own
files, following the pattern test_<concept>.cpp + test_<concept>_constexpr.cpp
