+++ test/test_group_event_loop_invalid_type_fail.cpp
@@ -10,7 +10,7 @@ struct NotAGroup
 struct DummyReceiver
 {
   using receives = ev_loop::type_list<int>;
-  void on(int) {}
+  void on(int /*unused*/) {}
 };
 
 // This should fail: NotAGroup is not a ThreadGroup or ExternalGroup
