+++ test/test_group_event_loop_no_groups_fail.cpp
