+++ test/test_group_event_loop_only_external_fail.cpp
